requirements.txt
dj-database-url==0.4.1
Django==1.11.1
google-auth==1.0.1
gunicorn==19.6.0
gspread==0.6.2
psycopg2==2.6.2
whitenoise==3.2
